// frontend/web/src/services/API/proserve-wb-publishing-api/models/Ami.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Publishing API
 * API in charge of publishing workbench products
 *
 * The version of the OpenAPI document: 2023-01-26T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Ami
 */
export interface Ami {
    /**
     * Unique ID of the AMI
     * @type {string}
     * @memberof Ami
     */
    amiId: string;
    /**
     * Name of the AMI
     * @type {string}
     * @memberof Ami
     */
    amiName: string;
    /**
     * Description of the AMI
     * @type {string}
     * @memberof Ami
     */
    amiDescription?: string;
    /**
     * Creation date of the AMI
     * @type {string}
     * @memberof Ami
     */
    createDate: string;
    /**
     * Last update date of the AMI
     * @type {string}
     * @memberof Ami
     */
    lastUpdateDate: string;
}

/**
 * Check if a given object implements the Ami interface.
 */
export function instanceOfAmi(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "amiId" in value;
    isInstance = isInstance && "amiName" in value;
    isInstance = isInstance && "createDate" in value;
    isInstance = isInstance && "lastUpdateDate" in value;

    return isInstance;
}

export function AmiFromJSON(json: any): Ami {
    return AmiFromJSONTyped(json, false);
}

export function AmiFromJSONTyped(json: any, ignoreDiscriminator: boolean): Ami {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'amiId': json['amiId'],
        'amiName': json['amiName'],
        'amiDescription': !exists(json, 'amiDescription') ? undefined : json['amiDescription'],
        'createDate': json['createDate'],
        'lastUpdateDate': json['lastUpdateDate'],
    };
}

export function AmiToJSON(value?: Ami | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'amiId': value.amiId,
        'amiName': value.amiName,
        'amiDescription': value.amiDescription,
        'createDate': value.createDate,
        'lastUpdateDate': value.lastUpdateDate,
    };
}